requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2.14
djangorestframework==3.15.2
Pillow==9.4.0
python-decouple==3.7
pytz==2022.7.1
sqlparse==0.4.3